Split handler
